requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
requests
pyyaml
google-api-python-client